[syntax] move syntaxw to syntax
except keep original pretty test
