+++ src/syntax/ast.rs
@@ -1,620 +1,462 @@
+///! ---------------------------------------------------------------------------------
+///! This code is auto generated by a tool $repo/scripts/ast.py
+///! Changes may cause incorrect behavior and will be lost if the code is regenerated.
+///! ---------------------------------------------------------------------------------
 
 use crate::source::{Span, IsId, IdSpan, FileId};
 use crate::lexical::{Separator, Keyword, Numeric};
+use crate::common::arena::{Arena, Index, Slice};
 
+#[cfg(test)]
+mod cmp;
+mod new;
+mod span;
+mod ugly;
+mod visit;
+mod noauto;
+mod pretty;
 mod profile;
-pub use profile::MemoryProfiler;
 
-// naming convension:
-// - first field is `span` and represents span for complete node
-// - Separator field is called `op`, its span is called `op_span`
-// - bracket pair span field is called `quote_span`
-// - get span methods for abc types are called `span`, e.g. `expr.span()`
-// - LabelDef field should be called `label`
-// - IsId from Token::Ident is normally called `name` type IdSpan
-// - use `parameters` not `params`
-// - concrete expression types should be called `XXXExpr`
-
-macro_rules! define_abc {
-    ($(#[$attr:meta])* $vis:vis enum $name:ident { $($variant:ident($ty:ident),)+ }) => (
-$(#[$attr])*
-$vis enum $name {
-    $($variant($ty),)+
+// put related methods in one module, because they are short
+// and may conflict with other short names if included by ast::*, use like 
+// ```
+// use ...ast::*;
+// ...
+// asti::Eq::eq(actual, expect);
+// (asti::debug(actual), asti::debug(expect));
+// ```
+pub mod asti {
+    #[cfg(test)]
+    pub use super::cmp::Eq;
+    pub use super::ugly::debug;
+    pub use super::pretty::{display, display1};
+    pub use super::profile::MemoryProfiler;
 }
+pub use span::WithSpan;
+// lit value can be regarded as a node type
+pub use noauto::LitValue;
+// natually extend to Arena when use ast::*
+pub use new::EmplaceConcrete;
 
-$(impl From<$ty> for $name {
-    fn from(v: $ty) -> $name {
-        $name::$variant(v)
-    }
-})+
-
-impl $name {
-    pub fn span(&self) -> Span {
-        match self {
-            $( Self::$variant(v) => v.span, )+
-        }
-    }
-}
-    );
-}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ArrayExpr {
-    pub span: Span, // bracket span
-    pub items: Vec<Expr>,
-}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ArrayIndexExpr {
+// AUTOGEN
+pub struct ArrayExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
-    pub parameters: Vec<Expr>,
-    pub quote_span: Span, // bracket span
+    pub items: Slice<'a, Expr<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ArrayType {
+pub struct ArrayIndexExpr<'a> {
     pub span: Span,
-    pub base: Box<TypeRef>,
-    pub size: Expr,
+    pub base: Expr<'a>,
+    pub parameters: Slice<'a, Expr<'a>>,
+    pub quote_span: Span,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct AssignExprStatement {
-    pub span: Span, // left.span + right.span
-    pub left: Expr,
-    pub right: Expr,
+pub struct ArrayType<'a> {
+    pub span: Span,
+    pub base: TypeRef<'a>,
+    pub size: Expr<'a>,
+}
+
+pub struct AssignExprStatement<'a> {
+    pub span: Span,
+    pub left: Expr<'a>,
+    pub right: Expr<'a>,
     pub op: Separator,
     pub op_span: Span,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct BinaryExpr {
+pub struct BinaryExpr<'a> {
     pub span: Span,
-    pub left: Box<Expr>,
-    pub right: Box<Expr>,
+    pub left: Expr<'a>,
+    pub right: Expr<'a>,
     pub op: Separator,
     pub op_span: Span,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct Block {
+pub struct Block<'a> {
     pub span: Span,
-    pub items: Vec<Statement>,
+    pub items: Slice<'a, Statement<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct BlockStatement {
+pub struct BlockStatement<'a> {
     pub span: Span,
     pub label: Option<IdSpan>,
-    pub body: Block,
+    pub body: Index<'a, Block<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct BreakStatement {
     pub span: Span,
     pub label: Option<IdSpan>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct CallExpr {
+pub struct CallExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
-    pub parameters: Vec<Expr>,
+    pub base: Expr<'a>,
     pub quote_span: Span,
+    pub parameters: Slice<'a, Expr<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct CastSegment {
+pub struct CastSegment<'a> {
     pub span: Span,
-    pub left: TypeRef,
-    pub right: TypeRef,
+    pub left: TypeRef<'a>,
+    pub right: TypeRef<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ClassDef {
+pub struct ClassDef<'a> {
     pub span: Span,
-    pub name: GenericName,
+    pub name: Index<'a, GenericName<'a>>,
     pub quote_span: Span,
-    pub items: Vec<Item>,
+    pub items: Slice<'a, Item<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct ContinueStatement {
     pub span: Span,
     pub label: Option<IdSpan>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ElseClause {
+pub struct ElseClause<'a> {
     pub span: Span,
-    pub body: Block,
+    pub body: Index<'a, Block<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct EnumDef {
+pub struct EnumDef<'a> {
     pub span: Span,
     pub name: IdSpan,
-    pub base_type: Option<PrimitiveType>,
+    pub base_type: Option<Index<'a, PrimitiveType>>,
     pub quote_span: Span,
-    pub variants: Vec<EnumDefVariant>,
+    pub variants: Slice<'a, Index<'a, EnumDefVariant<'a>>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct EnumDefVariant {
+pub struct EnumDefVariant<'a> {
     pub span: Span,
     pub name: IdSpan,
-    pub value: Option<Expr>,
+    pub value: Option<Expr<'a>>,
 }
 
-define_abc! {
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum Expr {
-    Lit(LitExpr),
-    Path(Path),
-    Paren(ParenExpr),
-    Tuple(TupleExpr),
-    Array(ArrayExpr),
-    Call(CallExpr),
-    ArrayIndex(ArrayIndexExpr),
-    TupleIndex(TupleIndexExpr),
-    Member(MemberExpr),
-    Object(ObjectExpr),
-    Unary(UnaryExpr),
-    Binary(BinaryExpr),
-    RangeBoth(RangeBothExpr),
-    RangeFull(RangeFullExpr),
-    RangeLeft(RangeLeftExpr),
-    RangeRight(RangeRightExpr),
-}}
-
-impl Expr {
-    pub fn dummy() -> Expr { 
-        Expr::Lit(LitExpr{ value: LitValue::Num(Numeric::I32(0)), span: Span::new(0, 0) }) 
-    }
+pub enum Expr<'a> {
+    Lit(Index<'a, LitExpr>),
+    Path(Index<'a, Path<'a>>),
+    Paren(Index<'a, ParenExpr<'a>>),
+    Tuple(Index<'a, TupleExpr<'a>>),
+    Array(Index<'a, ArrayExpr<'a>>),
+    Call(Index<'a, CallExpr<'a>>),
+    ArrayIndex(Index<'a, ArrayIndexExpr<'a>>),
+    TupleIndex(Index<'a, TupleIndexExpr<'a>>),
+    Member(Index<'a, MemberExpr<'a>>),
+    Object(Index<'a, ObjectExpr<'a>>),
+    Unary(Index<'a, UnaryExpr<'a>>),
+    Binary(Index<'a, BinaryExpr<'a>>),
+    RangeBoth(Index<'a, RangeBothExpr<'a>>),
+    RangeFull(Index<'a, RangeFullExpr>),
+    RangeLeft(Index<'a, RangeLeftExpr<'a>>),
+    RangeRight(Index<'a, RangeRightExpr<'a>>),
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct FieldDef {
+pub struct FieldDef<'a> {
     pub span: Span,
     pub name: IdSpan,
     pub colon_span: Span,
-    pub r#type: TypeRef,
+    pub r#type: TypeRef<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct FnDef {
-    pub span: Span, // fn_span + body_span
-    pub name: GenericName,
-    pub quote_span: Span, // parameter list quote
-    pub parameters: Vec<FnDefParameter>,
-    pub ret_type: Option<TypeRef>,
-    pub wheres: Vec<WhereClause>,
-    pub body: Option<Block>, // no body is allowed for intrinsic, extern and class associate methods
-}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct FnDefParameter {
-    pub span: Span, // name.span + r#type.span
-    pub name: IdSpan,
-    pub r#type: TypeRef,
-}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct FnType {
+pub struct FnDef<'a> {
     pub span: Span,
-    pub quote_span: Span, // parameter list quote span
-    pub parameters: Vec<FnTypeParameter>,
-    pub ret_type: Option<Box<TypeRef>>,
+    pub name: Index<'a, GenericName<'a>>,
+    pub quote_span: Span,
+    pub parameters: Slice<'a, Index<'a, FnDefParameter<'a>>>,
+    pub ret_type: Option<TypeRef<'a>>,
+    pub wheres: Slice<'a, Index<'a, WhereClause<'a>>>,
+    pub body: Option<Index<'a, Block<'a>>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct FnTypeParameter {
+pub struct FnDefParameter<'a> {
+    pub span: Span,
+    pub name: IdSpan,
+    pub r#type: TypeRef<'a>,
+}
+
+pub struct FnType<'a> {
+    pub span: Span,
+    pub quote_span: Span,
+    pub parameters: Slice<'a, Index<'a, FnTypeParameter<'a>>>,
+    pub ret_type: Option<TypeRef<'a>>,
+}
+
+pub struct FnTypeParameter<'a> {
     pub span: Span,
     pub name: Option<IdSpan>,
-    pub r#type: TypeRef,
+    pub r#type: TypeRef<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ForStatement {
+pub struct ForStatement<'a> {
     pub span: Span,
     pub label: Option<IdSpan>,
     pub iter_name: IdSpan,
-    pub iter_expr: Expr,
-    pub body: Block,
+    pub iter_expr: Expr<'a>,
+    pub body: Index<'a, Block<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct GenericName {
+pub struct GenericName<'a> {
     pub span: Span,
     pub base: IdSpan,
     pub quote_span: Span,
-    pub parameters: Vec<GenericParameter>,
+    pub parameters: Slice<'a, Index<'a, GenericParameter>>,
 }
 
-// currently this is only one identifier, but it's hard to pretty print vec<idspan> in one line
-// and future this may include where clause, so it is separate node
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct GenericParameter {
     pub span: Span,
     pub name: IdSpan,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct GenericSegment {
-    pub span: Span, 
-    pub base: IdSpan, 
-    pub parameters: TypeList,
+pub struct GenericSegment<'a> {
+    pub span: Span,
+    pub base: IdSpan,
+    pub parameters: Index<'a, TypeList<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct IfClause {
+pub struct IfClause<'a> {
     pub span: Span,
-    pub condition: Expr,
-    pub body: Block,
+    pub condition: Expr<'a>,
+    pub body: Index<'a, Block<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct IfStatement {
+pub struct IfStatement<'a> {
     pub span: Span,
-    pub if_clause: IfClause,
-    pub elseif_clauses: Vec<IfClause>,
-    pub else_clause: Option<ElseClause>,
+    pub if_clause: Index<'a, IfClause<'a>>,
+    pub elseif_clauses: Slice<'a, Index<'a, IfClause<'a>>>,
+    pub else_clause: Option<Index<'a, ElseClause<'a>>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct Implementation {
+pub struct Implementation<'a> {
     pub span: Span,
-    pub parameters: Vec<GenericParameter>,
-    pub class: Option<TypeRef>,
-    pub r#type: TypeRef,
-    pub wheres: Vec<WhereClause>,
+    pub parameters: Slice<'a, Index<'a, GenericParameter>>,
+    pub class: Option<TypeRef<'a>>,
+    pub r#type: TypeRef<'a>,
+    pub wheres: Slice<'a, Index<'a, WhereClause<'a>>>,
     pub quote_span: Span,
-    pub items: Vec<Item>,
+    pub items: Slice<'a, Item<'a>>,
 }
 
-define_abc! {
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum Item {
-    Struct(StructDef),
-    Enum(EnumDef),
-    Fn(FnDef),
-    Impl(Implementation),
-    Type(TypeDef),
-    Class(ClassDef),
-    Block(BlockStatement),
-    SimpleExpr(SimpleExprStatement),
-    AssignExpr(AssignExprStatement),
-    For(ForStatement),
-    If(IfStatement),
-    Loop(LoopStatement),
-    VarDecl(VarDeclStatement),
-    While(WhileStatement),
-    Use(UseStatement),
-    Import(ModuleStatement),
-}}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum LitValue {
-    Unit,
-    Bool(bool),
-    Char(char),
-    Str(IsId),
-    Num(Numeric),
+pub enum Item<'a> {
+    Struct(Index<'a, StructDef<'a>>),
+    Enum(Index<'a, EnumDef<'a>>),
+    Fn(Index<'a, FnDef<'a>>),
+    Impl(Index<'a, Implementation<'a>>),
+    Type(Index<'a, TypeDef<'a>>),
+    Class(Index<'a, ClassDef<'a>>),
+    Block(Index<'a, BlockStatement<'a>>),
+    SimpleExpr(Index<'a, SimpleExprStatement<'a>>),
+    AssignExpr(Index<'a, AssignExprStatement<'a>>),
+    For(Index<'a, ForStatement<'a>>),
+    If(Index<'a, IfStatement<'a>>),
+    Loop(Index<'a, LoopStatement<'a>>),
+    VarDecl(Index<'a, VarDeclStatement<'a>>),
+    While(Index<'a, WhileStatement<'a>>),
+    Use(Index<'a, UseStatement<'a>>),
+    Import(Index<'a, ModuleStatement>),
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct LitExpr {
     pub span: Span,
     pub value: LitValue,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct LoopStatement {
+pub struct LoopStatement<'a> {
     pub span: Span,
     pub label: Option<IdSpan>,
-    pub body: Block,
+    pub body: Index<'a, Block<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct MemberExpr {
+pub struct MemberExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
-    pub op_span: Span, // dot span
+    pub base: Expr<'a>,
+    pub op_span: Span,
     pub name: IdSpan,
-    pub parameters: Option<TypeList>,
+    pub parameters: Option<Index<'a, TypeList<'a>>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct Module {
+pub struct Module<'a> {
     pub file: FileId,
-    pub items: Vec<Item>,
+    pub items: Slice<'a, Item<'a>>,
 }
 
-// Attention: Clone for driver to store copy of import requests
-#[derive(Debug, Clone)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct ModuleStatement {
     pub span: Span,
     pub name: IdSpan,
     pub path: Option<IdSpan>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ObjectExpr {
+pub struct ObjectExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
+    pub base: Expr<'a>,
     pub quote_span: Span,
-    pub fields: Vec<ObjectExprField>,
+    pub fields: Slice<'a, Index<'a, ObjectExprField<'a>>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ObjectExprField {
+pub struct ObjectExprField<'a> {
     pub span: Span,
     pub name: IdSpan,
-    pub value: Expr,
+    pub value: Expr<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ParenExpr {
+pub struct ParenExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
+    pub base: Expr<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct Path {
+pub struct Path<'a> {
     pub span: Span,
-    pub segments: Vec<PathSegment>,
+    pub segments: Slice<'a, PathSegment<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum PathSegment {
-    // global is a mark to indicate path starts with namespace separator, 
-    // it does not have span because it logically located at the empty place before the namespace sparator
+pub enum PathSegment<'a> {
     Global,
-    Simple(SimpleSegment),
-    Cast(CastSegment),
-    Generic(GenericSegment),
+    Simple(Index<'a, SimpleSegment>),
+    Cast(Index<'a, CastSegment<'a>>),
+    Generic(Index<'a, GenericSegment<'a>>),
 }
 
-impl PathSegment {
-    pub fn span(&self) -> Span {
-        match self {
-            PathSegment::Global => Span::new(0, 0),
-            PathSegment::Simple(v) => v.span,
-            PathSegment::Cast(v) => v.span,
-            PathSegment::Generic(v) => v.span,
-        }
-    }
-}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct PrimitiveType {
     pub span: Span,
     pub base: Keyword,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct RangeBothExpr {
+pub struct RangeBothExpr<'a> {
     pub span: Span,
-    pub left: Box<Expr>,
+    pub left: Expr<'a>,
     pub op_span: Span,
-    pub right: Box<Expr>,
+    pub right: Expr<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct RangeFullExpr {
     pub span: Span,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct RangeLeftExpr {
+pub struct RangeLeftExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
+    pub base: Expr<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct RangeRightExpr {
+pub struct RangeRightExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
+    pub base: Expr<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct RefType {
+pub struct RefType<'a> {
     pub span: Span,
-    pub base: Box<TypeRef>,
+    pub base: TypeRef<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct ReturnStatement {
+pub struct ReturnStatement<'a> {
     pub span: Span,
-    pub value: Option<Expr>,
+    pub value: Option<Expr<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct SimpleExprStatement {
-    pub span: Span, // expr.span + semicolon_span
-    pub expr: Expr, 
+pub struct SimpleExprStatement<'a> {
+    pub span: Span,
+    pub expr: Expr<'a>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
 pub struct SimpleSegment {
     pub span: Span,
     pub name: IsId,
 }
 
-define_abc! {
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum Statement {
-    Struct(StructDef),
-    Enum(EnumDef),
-    Fn(FnDef),
-    Impl(Implementation),
-    Type(TypeDef),
-    Class(ClassDef),
-    Block(BlockStatement),
-    Break(BreakStatement),
-    Continue(ContinueStatement),
-    SimpleExpr(SimpleExprStatement),
-    AssignExpr(AssignExprStatement),
-    For(ForStatement),
-    If(IfStatement),
-    Loop(LoopStatement),
-    Return(ReturnStatement),
-    VarDecl(VarDeclStatement),
-    While(WhileStatement),
-    Use(UseStatement),
-}}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct StructDef {
-    pub span: Span,
-    pub name: GenericName,
-    pub fields: Vec<FieldDef>,
+pub enum Statement<'a> {
+    Struct(Index<'a, StructDef<'a>>),
+    Enum(Index<'a, EnumDef<'a>>),
+    Fn(Index<'a, FnDef<'a>>),
+    Impl(Index<'a, Implementation<'a>>),
+    Type(Index<'a, TypeDef<'a>>),
+    Class(Index<'a, ClassDef<'a>>),
+    Block(Index<'a, BlockStatement<'a>>),
+    Break(Index<'a, BreakStatement>),
+    Continue(Index<'a, ContinueStatement>),
+    SimpleExpr(Index<'a, SimpleExprStatement<'a>>),
+    AssignExpr(Index<'a, AssignExprStatement<'a>>),
+    For(Index<'a, ForStatement<'a>>),
+    If(Index<'a, IfStatement<'a>>),
+    Loop(Index<'a, LoopStatement<'a>>),
+    Return(Index<'a, ReturnStatement<'a>>),
+    VarDecl(Index<'a, VarDeclStatement<'a>>),
+    While(Index<'a, WhileStatement<'a>>),
+    Use(Index<'a, UseStatement<'a>>),
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct TupleExpr {
+pub struct StructDef<'a> {
     pub span: Span,
-    pub items: Vec<Expr>,
+    pub name: Index<'a, GenericName<'a>>,
+    pub fields: Slice<'a, Index<'a, FieldDef<'a>>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct TupleIndexExpr {
+pub struct TupleExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>,
-    pub op_span: Span, // dot span
+    pub items: Slice<'a, Expr<'a>>,
+}
+
+pub struct TupleIndexExpr<'a> {
+    pub span: Span,
+    pub base: Expr<'a>,
+    pub op_span: Span,
     pub value: i32,
     pub value_span: Span,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct TupleType {
+pub struct TupleType<'a> {
     pub span: Span,
-    // not TypeList, TypeList is angle bracket quoted, tuple type is paren quoted and empty allowed
-    pub parameters: Vec<TypeRef>,
+    pub parameters: Slice<'a, TypeRef<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct TypeDef {
+pub struct TypeDef<'a> {
     pub span: Span,
-    pub name: GenericName,
-    pub from: Option<TypeRef>,
+    pub name: Index<'a, GenericName<'a>>,
+    pub from: Option<TypeRef<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct TypeList {
+pub struct TypeList<'a> {
     pub span: Span,
-    pub items: Vec<TypeRef>,
+    pub items: Slice<'a, TypeRef<'a>>,
 }
 
-define_abc! {
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum TypeRef {
-    Array(ArrayType),
-    Fn(FnType),
-    Path(Path),
-    Primitive(PrimitiveType),
-    Ref(RefType),
-    Tuple(TupleType),
-}}
+pub enum TypeRef<'a> {
+    Array(Index<'a, ArrayType<'a>>),
+    Fn(Index<'a, FnType<'a>>),
+    Path(Index<'a, Path<'a>>),
+    Primitive(Index<'a, PrimitiveType>),
+    Ref(Index<'a, RefType<'a>>),
+    Tuple(Index<'a, TupleType<'a>>),
+}
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct UnaryExpr {
+pub struct UnaryExpr<'a> {
     pub span: Span,
-    pub base: Box<Expr>, 
-    pub op: Separator, 
+    pub base: Expr<'a>,
+    pub op: Separator,
     pub op_span: Span,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct UseStatement {
+pub struct UseStatement<'a> {
     pub span: Span,
-    pub path: Path,
+    pub path: Index<'a, Path<'a>>,
     pub alias: Option<IdSpan>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct VarDeclStatement {
+pub struct VarDeclStatement<'a> {
     pub span: Span,
     pub r#const: bool,
     pub name: IdSpan,
-    pub r#type: Option<TypeRef>,
-    pub init_value: Option<Expr>,
+    pub r#type: Option<TypeRef<'a>>,
+    pub init_value: Option<Expr<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct WhereClause {
+pub struct WhereClause<'a> {
     pub span: Span,
     pub name: IdSpan,
-    pub constraints: Vec<TypeRef>,
+    pub constraints: Slice<'a, TypeRef<'a>>,
 }
 
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct WhileStatement {
+pub struct WhileStatement<'a> {
     pub span: Span,
     pub label: Option<IdSpan>,
-    pub condition: Expr,
-    pub body: Block,
+    pub condition: Expr<'a>,
+    pub body: Index<'a, Block<'a>>,
 }
