+++ src/syntax/ast/profile.rs
@@ -1,378 +1,352 @@
+///! ---------------------------------------------------------------------------------
+///! This code is auto generated by a tool $repo/scripts/ast.py
+///! Changes may cause incorrect behavior and will be lost if the code is regenerated.
+///! ---------------------------------------------------------------------------------
+
 ///! ast memory usage profiler
 
-use std::collections::HashMap;
-use crate::common::arena::{Index, Slice};
-use super::super::visit::{Visitor, Node};
+use std::mem::size_of;
+use crate::common::arena::Arena;
+use super::super::visit::{Visitor, Visit};
+use super::noauto::EmptyResult;
 use super::*;
 
-#[derive(Default)]
-struct MemoryUsage {
+#[derive(Clone, Copy)]
+struct Entry {
     count: usize,
-    // in bytes, same as result of mem::size_of
-    // for normal node,
-    //   size is size of all owned fields, which contains
-    //     all "terminal" fields (include IsId, Span, IdSpan, Separator, Keyword, LitValue)
-    //     and box and vec's shallow size (sizeof!(Box<T>) (or 1 size_t) and sizeof!(Vec<T>>) (or 3 size_t))
-    //     and option tag (e.g. sizeof!(Option<Expr>) - sizeof!(Expr)), because rust is not smart/aggresive enough to fold union tags
-    //     without directly owned node (e.g. ForStatement::condition)
-    // for abc node,
-    //   size is tag + pad for largest variant, which is sizeof!(Self) - sizeof!(CurrentVariant)
-    total_size: usize,
-    // in bytes, same as result of mem::size_of
-    // for normal node,
-    //   theoretical size is size of all owned fields, with help of unified object pool, which contains
-    //     all "terminal" fields
-    //     for Box and directly owned node, u32 for object index
-    //     for Vec, u32 for vec length and u32 for each item's object index
-    //     for Option, plus u32 for option tag
-    // for abc node,
-    //   theoretical size is u32, avoid this padding is one of the major targets of unified object pool
-    total_theoretical_size: usize,
+    size: usize,
 }
 
-macro_rules! sizeof {
-    ($ty:ty) => (std::mem::size_of::<$ty>());
-}
+impl Entry {
 
-impl MemoryUsage {
-
-    fn count_self(&mut self) -> &mut Self {
+    fn count(&mut self, size: usize) {
         self.count += 1;
-        self
+        self.size += size;
     }
-    fn count_impl(&mut self, size: usize, theoretical_size: usize) -> &mut Self {
-        self.total_size += size;
-        self.total_theoretical_size += theoretical_size;
-        self
-    }
-
-    // count terminal, used frequently so name is short
-    // .label, option<isid> etc. all use this
-    fn count_t<T>(&mut self, _: &T) -> &mut Self {
-        self.count_impl(sizeof!(T), sizeof!(T))
-    }
-    // option node, not option<idspan> like
-    fn count_option<N: Node>(&mut self, _: &Option<N>) -> &mut Self {
-        self.count_impl(sizeof!(Option<N>) - sizeof!(N), sizeof!(Option<Index<()>>))
-    }
-    fn count_option_box_owned<T>(&mut self, _: &Option<Box<T>>) -> &mut Self {
-        self.count_impl(sizeof!(Option<Box<T>>), sizeof!(Option<Index<()>>))
-    }
-    fn count_box_owned<T>(&mut self, _: &Box<T>) -> &mut Self {
-        self.count_impl(sizeof!(Box<T>), sizeof!(Index<()>))
-    }
-    // count directly owned node
-    fn count_owned<N: Node>(&mut self, _: &N) -> &mut Self {
-        self.count_impl(0, sizeof!(Index<()>))
-    }
-    fn count_vec<T>(&mut self, _: &Vec<T>) -> &mut Self {
-        self.count_impl(sizeof!(Vec<T>), sizeof!(Slice<()>))
-    }
-}
-
-pub struct MemoryProfiler {
-    // key is node name
-    //   for normal node, key is node name
-    //   for abc node, will use key for each variant (e.g. expr (array-expr), expr (tuple-expr))
-    u: HashMap<&'static str, MemoryUsage>,
 }
 
 impl MemoryProfiler {
 
-    pub fn new() -> Self {
-        Self{ u: HashMap::new() }
-    }
-
-    pub fn profile<N: Node>(&mut self, node: &N) {
-        let _ = node.accept(self);
-    }
-
     pub fn dump(&self, output: &mut impl std::io::Write) {
-        let total_total_size: usize = self.u.values().map(|v| v.total_size).sum();
-        let total_total_theoretical_size: usize = self.u.values().map(|v| v.total_theoretical_size).sum();
-        let mut ordered_items = self.u.iter().collect::<Vec<_>>();
-        ordered_items.sort_unstable_by(|a, b| a.1.total_size.cmp(&b.1.total_size).then_with(|| a.0.cmp(&b.0)).reverse());
-        writeln!(output, "total size {total_total_size} > {total_total_theoretical_size}").unwrap();
-        for (&key, MemoryUsage{ count, total_size, total_theoretical_size }) in ordered_items {
-            let header = format!("[{key}({count})]");
-            let total_size_percent = *total_size as f64 * 100f64 / total_total_size as f64;
-            let total_theoretical_size_percent = *total_theoretical_size as f64 * 100f64 / total_total_theoretical_size as f64;
-            writeln!(output, "{header:>28} {total_size} ({total_size_percent:.2}%) > {total_theoretical_size} ({total_theoretical_size_percent:.2}%)").unwrap();
+
+        let total_count: usize = self.v.iter().map(|e| e.count).sum();
+        let total_size: usize = self.v.iter().map(|e| e.size).sum();
+        writeln!(output, "[all({total_count})] {total_size}").unwrap();
+
+        let mut ordered_items = self.v.iter()
+            .enumerate().map(|(i, e)| (NAMES[i], e.count, e.size)).collect::<Vec<_>>();
+        ordered_items.sort_unstable_by(|a, b| a.2.cmp(&b.2).then_with(|| a.0.cmp(&b.0)).reverse());
+        for (key, count, size) in ordered_items {
+            if count > 0 {
+                let header = format!("[{key}({count})]");
+                let size_percent = size as f64 * 100f64 / total_size as f64;
+                writeln!(output, "{header} {size} ({size_percent:.2}%)").unwrap();
+            }
         }
     }
 }
 
-macro_rules! impl_visit {
-    ($ty:ty, $fn:ident, $name:literal $(,)? $($countf:ident(.$field:ident),)+) => (
-        fn $fn(&mut self, node: &$ty) -> Result<(), ()> {
-            self.u.entry($name).or_default().count_self()$(.$countf(&node.$field))+;
-            node.walk(self)
-        }
-    );
-    (enum $ty:ty, $fn:ident, $name:literal $(,)? $($variant:ident($subty:ty), $subname:literal,)+) => (
-        fn $fn(&mut self, node: &$ty) -> Result<(), ()> {
-            use $ty::*;
-            match node {
-            $(
-                $variant(_) => self.u.entry(concat!($name, "(", $subname, ")")) // u32 for enum tag and min align is 4 
-                    .or_default().count_self().count_impl(sizeof!($ty), sizeof!(u32) + sizeof!(Index<()>)),
-            )+
-            };
-            node.walk(self)
-        }
-    );
+// AUTOGEN
+// use static name list because they are known at code generation time
+static NAMES: &[&str] = &[
+    "array-expr",
+    "array-index-expr",
+    "array-type",
+    "assign-expr-stmt",
+    "binary-expr",
+    "block",
+    "block-stmt",
+    "break-stmt",
+    "call-expr",
+    "cast-segment",
+    "class-def",
+    "continue-stmt",
+    "else-clause",
+    "enum-def",
+    "enum-def-variant",
+    "field-def",
+    "fn-def",
+    "fn-def-parameter",
+    "fn-type",
+    "fn-type-parameter",
+    "for-stmt",
+    "generic-name",
+    "generic-parameter",
+    "generic-segment",
+    "if-clause",
+    "if-stmt",
+    "impl-block",
+    "lit-expr",
+    "loop-stmt",
+    "member-expr",
+    "module",
+    "module-stmt",
+    "object-expr",
+    "object-expr-field",
+    "paren-expr",
+    "path",
+    "primitive-type",
+    "range-both-expr",
+    "range-full-expr",
+    "range-left-expr",
+    "range-right-expr",
+    "ref-type",
+    "ret-stmt",
+    "simple-expr-stmt",
+    "simple-segment",
+    "struct-def",
+    "tuple-expr",
+    "tuple-index-expr",
+    "tuple-type",
+    "type-def",
+    "type-list",
+    "unary-expr",
+    "use-stmt",
+    "var-decl-stmt",
+    "where-clause",
+    "while-stmt",
+];
+
+pub struct MemoryProfiler {
+    v: [Entry; 56],
+}
+
+impl MemoryProfiler {
+    pub fn new() -> Self {
+        Self{ v: [Entry{ count: 0, size: 0 }; 56] }
+    }
+    pub fn coverage(&self) -> (usize, usize, String) {
+        (self.v.iter().filter(|e| e.count > 0).count(), self.v.len(),
+            self.v.iter().enumerate().filter(|(_, e)| e.count == 0).map(|(i, _)| NAMES[i]).collect::<Vec<_>>().join(", "))
+    }
 }
 
 impl Visitor for MemoryProfiler {
+    type Result = EmptyResult;
+    // enum types are all default here because they literal does not use any memory on its own in this arena refactor
 
-    impl_visit!{ ArrayExpr, visit_array_expr, "array-expr",
-        count_t(.span), count_vec(.items),
+    fn visit_array_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ArrayExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[0].count(size_of::<ArrayExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ArrayIndexExpr, visit_array_index_expr, "array-index-expr",
-        count_t(.span), count_box_owned(.base), count_vec(.parameters), count_t(.quote_span),
+    fn visit_array_index_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ArrayIndexExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[1].count(size_of::<ArrayIndexExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ArrayType, visit_array_type, "array-type",
-        count_t(.span), count_box_owned(.base), count_owned(.size),
+    fn visit_array_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ArrayType<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[2].count(size_of::<ArrayType>());
+        node.walk(arena, self)
     }
-    impl_visit!{ AssignExprStatement, visit_assign_expr_stmt, "assign-expr-stmt",
-        count_t(.span), count_owned(.left), count_owned(.right), count_t(.op), count_t(.op_span),
+    fn visit_assign_expr_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, AssignExprStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[3].count(size_of::<AssignExprStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ BinaryExpr, visit_binary_expr, "binary-expr",
-        count_t(.span), count_box_owned(.left), count_box_owned(.right), count_t(.op), count_t(.op_span),
+    fn visit_binary_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, BinaryExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[4].count(size_of::<BinaryExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ Block, visit_block, "block",
-        count_t(.span), count_vec(.items),
+    fn visit_block<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Block<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[5].count(size_of::<Block>());
+        node.walk(arena, self)
     }
-    impl_visit!{ BlockStatement, visit_block_stmt, "block-stmt",
-        count_t(.span), count_t(.label), count_owned(.body),
+    fn visit_block_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, BlockStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[6].count(size_of::<BlockStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ BreakStatement, visit_break_stmt, "break-stmt",
-        count_t(.span), count_t(.label),
+    fn visit_break_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, BreakStatement>, arena: &'a Arena) -> Self::Result {
+        self.v[7].count(size_of::<BreakStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ CallExpr, visit_call_expr, "call-expr",
-        count_t(.span), count_box_owned(.base), count_vec(.parameters), count_t(.quote_span),
+    fn visit_call_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, CallExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[8].count(size_of::<CallExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ CastSegment, visit_cast_segment, "cast-segment",
-        count_t(.span), count_owned(.left), count_owned(.right),
+    fn visit_cast_segment<'a, 'b: 'a>(&mut self, node: &'b Index<'a, CastSegment<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[9].count(size_of::<CastSegment>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ClassDef, visit_class_def, "class-def",
-        count_t(.span), count_owned(.name), count_t(.quote_span), count_vec(.items),
+    fn visit_class_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ClassDef<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[10].count(size_of::<ClassDef>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ContinueStatement, visit_continue_stmt, "continue-stmt",
-        count_t(.span), count_t(.label),
+    fn visit_continue_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ContinueStatement>, arena: &'a Arena) -> Self::Result {
+        self.v[11].count(size_of::<ContinueStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ElseClause, visit_else_clause, "else-clause",
-        count_t(.span), count_owned(.body),
+    fn visit_else_clause<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ElseClause<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[12].count(size_of::<ElseClause>());
+        node.walk(arena, self)
     }
-    impl_visit!{ EnumDef, visit_enum_def, "enum-def",
-        count_t(.span), count_t(.name), count_option(.base_type), count_t(.quote_span), count_vec(.variants),
+    fn visit_enum_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, EnumDef<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[13].count(size_of::<EnumDef>());
+        node.walk(arena, self)
     }
-    impl_visit!{ EnumDefVariant, visit_enum_def_variant, "enum-def-variant",
-        count_t(.span), count_t(.name), count_option(.value),
+    fn visit_enum_def_variant<'a, 'b: 'a>(&mut self, node: &'b Index<'a, EnumDefVariant<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[14].count(size_of::<EnumDefVariant>());
+        node.walk(arena, self)
     }
-    impl_visit!{ enum Expr, visit_expr, "expr",
-        Lit(LitExpr), "lit",
-        Path(super::Path), "path",
-        Paren(ParenExpr), "paren",
-        Tuple(TupleExpr), "tuple",
-        Array(ArrayExpr), "array",
-        Call(CallExpr), "call",
-        ArrayIndex(ArrayIndexExpr), "array-index",
-        TupleIndex(TupleIndexExpr), "tuple-index",
-        Member(MemberExpr), "member",
-        Object(ObjectExpr), "object",
-        Unary(UnaryExpr), "unary",
-        Binary(BinaryExpr), "binary",
-        RangeBoth(RangeBothExpr), "range-both",
-        RangeFull(RangeFullExpr), "range-full",
-        RangeLeft(RangeLeftExpr), "range-left",
-        RangeRight(RangeRightExpr), "range-right",
+    fn visit_field_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FieldDef<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[15].count(size_of::<FieldDef>());
+        node.walk(arena, self)
     }
-    impl_visit!{ FieldDef, visit_field_def, "field-def",
-        count_t(.span), count_t(.name), count_t(.colon_span), count_owned(.r#type),
+    fn visit_fn_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnDef<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[16].count(size_of::<FnDef>());
+        node.walk(arena, self)
     }
-    impl_visit!{ FnDef, visit_fn_def, "fn-def",
-        count_t(.span), count_owned(.name), count_t(.span), count_vec(.parameters), count_option(.ret_type), count_vec(.wheres), count_option(.body),
+    fn visit_fn_def_parameter<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnDefParameter<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[17].count(size_of::<FnDefParameter>());
+        node.walk(arena, self)
     }
-    impl_visit!{ FnDefParameter, visit_fn_def_parameter, "fn-def-parameter",
-        count_t(.span), count_t(.name), count_owned(.r#type),
+    fn visit_fn_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnType<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[18].count(size_of::<FnType>());
+        node.walk(arena, self)
     }
-    impl_visit!{ FnType, visit_fn_type, "fn-type",
-        count_t(.span), count_t(.quote_span), count_vec(.parameters), count_option_box_owned(.ret_type),
+    fn visit_fn_type_parameter<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnTypeParameter<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[19].count(size_of::<FnTypeParameter>());
+        node.walk(arena, self)
     }
-    impl_visit!{ FnTypeParameter, visit_fn_type_parameter, "fn-type-parameter",
-        count_t(.span), count_t(.name), count_owned(.r#type),
+    fn visit_for_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ForStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[20].count(size_of::<ForStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ForStatement, visit_for_stmt, "for-stmt",
-        count_t(.span), count_t(.label), count_t(.iter_name), count_owned(.iter_expr), count_owned(.body),
+    fn visit_generic_name<'a, 'b: 'a>(&mut self, node: &'b Index<'a, GenericName<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[21].count(size_of::<GenericName>());
+        node.walk(arena, self)
     }
-    impl_visit!{ GenericName, visit_generic_name, "generic-name",
-        count_t(.span), count_t(.base), count_t(.quote_span), count_vec(.parameters),
+    fn visit_generic_parameter<'a, 'b: 'a>(&mut self, node: &'b Index<'a, GenericParameter>, arena: &'a Arena) -> Self::Result {
+        self.v[22].count(size_of::<GenericParameter>());
+        node.walk(arena, self)
     }
-    impl_visit!{ GenericParameter, visit_generic_parameter, "generic-parameter",
-        count_t(.span), count_t(.name),
+    fn visit_generic_segment<'a, 'b: 'a>(&mut self, node: &'b Index<'a, GenericSegment<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[23].count(size_of::<GenericSegment>());
+        node.walk(arena, self)
     }
-    impl_visit!{ GenericSegment, visit_generic_segment, "generic-segment",
-        count_t(.span), count_t(.base), count_owned(.parameters),
+    fn visit_if_clause<'a, 'b: 'a>(&mut self, node: &'b Index<'a, IfClause<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[24].count(size_of::<IfClause>());
+        node.walk(arena, self)
     }
-    impl_visit!{ IfClause, visit_if_clause, "if-clause",
-        count_t(.span), count_owned(.condition), count_owned(.body),
+    fn visit_if_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, IfStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[25].count(size_of::<IfStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ IfStatement, visit_if_stmt, "if-stmt",
-        count_t(.span), count_owned(.if_clause), count_vec(.elseif_clauses), count_option(.else_clause),
+    fn visit_impl_block<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Implementation<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[26].count(size_of::<Implementation>());
+        node.walk(arena, self)
     }
-    impl_visit!{ Implementation, visit_impl, "impl-block",
-        count_t(.span), count_vec(.parameters), count_option(.class), count_owned(.r#type), count_vec(.wheres), count_t(.quote_span), count_vec(.items),
+    fn visit_lit_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, LitExpr>, arena: &'a Arena) -> Self::Result {
+        self.v[27].count(size_of::<LitExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ enum Item, visit_item, "item",
-        Struct(StructDef), "struct",
-        Enum(EnumDef), "enum",
-        Fn(FnDef), "fn",
-        Impl(Implementation), "impl",
-        Type(TypeDef), "type",
-        Class(ClassDef), "class",
-        Block(BlockStatement), "block",
-        SimpleExpr(SimpleExprStatement), "simple-expr",
-        AssignExpr(AssignExprStatement), "assign-expr",
-        For(ForStatement), "for",
-        If(IfStatement), "if",
-        Loop(LoopStatement), "loop",
-        VarDecl(VarDeclStatement), "var-decl",
-        While(WhileStatement), "while",
-        Use(UseStatement), "use",
-        Import(ModuleStatement), "import",
+    fn visit_loop_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, LoopStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[28].count(size_of::<LoopStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ LitExpr, visit_lit_expr, "lit-expr",
-        count_t(.span), count_t(.value),
+    fn visit_member_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, MemberExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[29].count(size_of::<MemberExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ LoopStatement, visit_loop_stmt, "loop-stmt",
-        count_t(.span), count_t(.label), count_owned(.body),
+    fn visit_module<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Module<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[30].count(size_of::<Module>());
+        node.walk(arena, self)
     }
-    impl_visit!{ MemberExpr, visit_member_expr, "member-expr",
-        count_t(.span), count_box_owned(.base), count_t(.op_span), count_t(.name), count_option(.parameters),
+    fn visit_module_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ModuleStatement>, arena: &'a Arena) -> Self::Result {
+        self.v[31].count(size_of::<ModuleStatement>());
+        node.walk(arena, self)
     }
-    // impl_visit!{ Module }
-    impl_visit!{ ModuleStatement, visit_module_stmt, "module-stmt",
-        count_t(.span), count_t(.name), count_t(.path),
+    fn visit_object_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ObjectExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[32].count(size_of::<ObjectExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ObjectExpr, visit_object_expr, "object-expr",
-        count_t(.span), count_box_owned(.base), count_t(.span), count_vec(.fields),
+    fn visit_object_expr_field<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ObjectExprField<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[33].count(size_of::<ObjectExprField>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ObjectExprField, visit_object_expr_field, "object-expr-field",
-        count_t(.span), count_t(.name), count_owned(.value),
+    fn visit_paren_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ParenExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[34].count(size_of::<ParenExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ParenExpr, visit_paren_expr, "paren-expr",
-        count_t(.span), count_box_owned(.base),
+    fn visit_path<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Path<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[35].count(size_of::<Path>());
+        node.walk(arena, self)
     }
-    impl_visit!{ Path, visit_path, "path",
-        count_t(.span), count_vec(.segments),
+    fn visit_primitive_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, PrimitiveType>, arena: &'a Arena) -> Self::Result {
+        self.v[36].count(size_of::<PrimitiveType>());
+        node.walk(arena, self)
     }
-
-    // this is combination of abc and normal
-    fn visit_path_segment(&mut self, node: &PathSegment) -> Result<(), ()> {
-        match node {
-            PathSegment::Global => {
-                self.u.entry("path-segment(global)").or_default().count_self().count_impl(sizeof!(PathSegment), sizeof!(u32) + sizeof!(Index<()>));
-            },
-            PathSegment::Simple(_) => {
-                self.u.entry("path-segment(simple)").or_default().count_self().count_impl(sizeof!(PathSegment), sizeof!(u32) + sizeof!(Index<()>));
-            },
-            PathSegment::Cast(_) => {
-                self.u.entry("path-segment(cast)").or_default().count_self().count_impl(sizeof!(PathSegment), sizeof!(u32) + sizeof!(Index<()>));
-            },
-            PathSegment::Generic(_) => {
-                self.u.entry("path-segment(generic)").or_default().count_self().count_impl(sizeof!(PathSegment), sizeof!(u32) + sizeof!(Index<()>));
-            },
-        }
-        node.walk(self)
+    fn visit_range_both_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeBothExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[37].count(size_of::<RangeBothExpr>());
+        node.walk(arena, self)
     }
-
-    impl_visit!{ PrimitiveType, visit_primitive_type, "primitive-type",
-        count_t(.span), count_t(.base),
+    fn visit_range_full_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeFullExpr>, arena: &'a Arena) -> Self::Result {
+        self.v[38].count(size_of::<RangeFullExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ RangeBothExpr, visit_range_both_expr, "range-both-expr",
-        count_t(.span), count_box_owned(.left), count_t(.op_span), count_box_owned(.right),
+    fn visit_range_left_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeLeftExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[39].count(size_of::<RangeLeftExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ RangeFullExpr, visit_range_full_expr, "range-full-expr",
-        count_t(.span),
+    fn visit_range_right_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeRightExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[40].count(size_of::<RangeRightExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ RangeLeftExpr, visit_range_left_expr, "range-left-expr",
-        count_t(.span), count_box_owned(.base),
+    fn visit_ref_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RefType<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[41].count(size_of::<RefType>());
+        node.walk(arena, self)
     }
-    impl_visit!{ RangeRightExpr, visit_range_right_expr, "range-right-expr",
-        count_t(.span), count_box_owned(.base),
+    fn visit_ret_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ReturnStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[42].count(size_of::<ReturnStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ RefType, visit_ref_type, "ref-type"
-        count_t(.span), count_box_owned(.base),
+    fn visit_simple_expr_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, SimpleExprStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[43].count(size_of::<SimpleExprStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ ReturnStatement, visit_ret_stmt, "ret-stmt"
-        count_t(.span), count_option(.value),
+    fn visit_simple_segment<'a, 'b: 'a>(&mut self, node: &'b Index<'a, SimpleSegment>, arena: &'a Arena) -> Self::Result {
+        self.v[44].count(size_of::<SimpleSegment>());
+        node.walk(arena, self)
     }
-    impl_visit!{ SimpleExprStatement, visit_simple_expr_stmt, "simple-expr-stmt"
-        count_t(.span), count_owned(.expr),
+    fn visit_struct_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, StructDef<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[45].count(size_of::<StructDef>());
+        node.walk(arena, self)
     }
-    impl_visit!{ SimpleSegment, visit_simple_segment, "simple-segment",
-        count_t(.span), count_t(.name),
+    fn visit_tuple_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TupleExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[46].count(size_of::<TupleExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ enum Statement, visit_stmt, "stmt" 
-        Struct(StructDef), "struct",
-        Enum(EnumDef), "enum",
-        Fn(FnDef), "fn",
-        Impl(Implementation), "impl",
-        Type(TypeDef), "type",
-        Class(ClassDef), "class",
-        Block(BlockStatement), "block",
-        Break(BreakStatement), "break",
-        Continue(ContinueStatement), "continue",
-        SimpleExpr(SimpleExprStatement), "simple-expr",
-        AssignExpr(AssignExprStatement), "assign-expr",
-        For(ForStatement), "for",
-        If(IfStatement), "if",
-        Loop(LoopStatement), "loop",
-        Return(ReturnStatement), "ret",
-        VarDecl(VarDeclStatement), "var-decl",
-        While(WhileStatement), "while",
-        Use(UseStatement), "use",
+    fn visit_tuple_index_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TupleIndexExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[47].count(size_of::<TupleIndexExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ StructDef, visit_struct_def, "struct-def"
-        count_t(.span), count_owned(.name), count_vec(.fields),
+    fn visit_tuple_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TupleType<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[48].count(size_of::<TupleType>());
+        node.walk(arena, self)
     }
-    impl_visit!{ TupleExpr, visit_tuple_expr, "tuple-expr"
-        count_t(.span), count_vec(.items),
+    fn visit_type_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TypeDef<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[49].count(size_of::<TypeDef>());
+        node.walk(arena, self)
     }
-    impl_visit!{ TupleIndexExpr, visit_tuple_index_expr, "tuple-index-expr"
-        count_t(.span), count_box_owned(.base), count_t(.op_span), count_t(.value),
+    fn visit_type_list<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TypeList<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[50].count(size_of::<TypeList>());
+        node.walk(arena, self)
     }
-    impl_visit!{ TupleType, visit_tuple_type, "tuple-type"
-        count_t(.span), count_vec(.parameters),
+    fn visit_unary_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, UnaryExpr<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[51].count(size_of::<UnaryExpr>());
+        node.walk(arena, self)
     }
-    impl_visit!{ TypeDef, visit_type_def, "type-def",
-        count_t(.span), count_owned(.name), count_option(.from),
+    fn visit_use_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, UseStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[52].count(size_of::<UseStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ TypeList, visit_type_list, "type-list"
-        count_t(.span), count_vec(.items),
+    fn visit_var_decl_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, VarDeclStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[53].count(size_of::<VarDeclStatement>());
+        node.walk(arena, self)
     }
-    impl_visit!{ enum TypeRef, visit_type_ref, "type-ref"
-        Array(ArrayType), "array",
-        Fn(FnType), "fn",
-        Path(super::Path), "path",
-        Primitive(PrimitiveType), "primitive",
-        Ref(RefType), "ref",
-        Tuple(TupleType), "tuple",
+    fn visit_where_clause<'a, 'b: 'a>(&mut self, node: &'b Index<'a, WhereClause<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[54].count(size_of::<WhereClause>());
+        node.walk(arena, self)
     }
-    impl_visit!{ UnaryExpr, visit_unary_expr, "unary-expr"
-        count_t(.span), count_box_owned(.base), count_t(.op), count_t(.op_span),
-    }
-    impl_visit!{ UseStatement, visit_use_stmt, "use-stmt"
-        count_t(.span), count_owned(.path), count_t(.alias),
-    }
-    impl_visit!{ VarDeclStatement, visit_var_decl, "var-decl"
-        count_t(.span), count_t(.r#const), count_t(.name), count_option(.r#type), count_option(.init_value),
-    }
-    impl_visit!{ WhereClause, visit_where_clause, "where-clause"
-        count_t(.span), count_t(.name), count_vec(.constraints),
-    }
-    impl_visit!{ WhileStatement, visit_while_stmt, "while-stmt"
-        count_t(.span), count_t(.label), count_owned(.condition), count_owned(.body),
+    fn visit_while_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, WhileStatement<'a>>, arena: &'a Arena) -> Self::Result {
+        self.v[55].count(size_of::<WhileStatement>());
+        node.walk(arena, self)
     }
 }
