+++ src/syntax/visit.rs
@@ -1,97 +1,105 @@
-///! syntax::visit: visitor and visitee traits
+///! ---------------------------------------------------------------------------------
+///! This code is auto generated by a tool $repo/scripts/ast.py
+///! Changes may cause incorrect behavior and will be lost if the code is regenerated.
+///! ---------------------------------------------------------------------------------
 
-use crate::source::SourceContext;
-use super::pretty::NodeDisplay;
-use super::ast::*;
-
-pub trait Node: Sized {
-
-    fn accept<T: Default, E, V: Visitor<T, E>>(&self, _visitor: &mut V) -> Result<T, E>;
-
-    fn walk<T: Default, E, V: Visitor<T, E>>(&self, _visitor: &mut V) -> Result<T, E> { 
-        Ok(Default::default())
-    }
-
-    // this should be some trait NodeDisplay: Node in theory but directly put this here is convenient
-    fn display<'n, 'scx, F>(&'n self, context: &'scx SourceContext<F>) -> NodeDisplay<'n, 'scx, Self, F> {
-        NodeDisplay{ node: self, context }
-    }
-}
+///! visitor and visitable trait
 
 // visitor visits node
-// - each visit method visits specified node type and do different things to collect information (not alter, at least this trait)
-//   after that, Node::walk is called to let node itself guide visitor to child nodes, this decouples visitor implementation with tree structure
-// - on the other hand, Node::accept implementation should call this node's matching visit_* method without dothing other things
-//   only then can you visit an arbitray 'impl Node' type, or 'dyn Node' if it is object safe (but not this Node)
+// - each visit method visits specific node type and do different things 
+//   to collect information (or change, by VisitorMut), after that, Visit::walk should be called to let
+//   node itself guide visitor to child nodes, this decouples visitor implementation with tree structure
+//   UPDATE: 
+//   - eq is not using visitor because it is comparing a pair of node, not visiting one node
+//   - ugly and pretty is not using walk because they need to display something between child nodes
+//   - so currently only profile is actually using walk because it is a simple count(size_of)
+// - on the other hand, Visit::accept implementation should call this node's matching visit method 
+//   without dothing any other things, only then can methods accepting arbitrary V: Visit perform visit operation
 // - according to this design
 //   - default visit implementation walks self
 //   - default walk implementation does nothing
 //   - no default accept implementation
-// - the return type is Result, because Node::walk normally need to call visit methods multiple times,
-//   so there should be some way to "aggregate" them, and Result (Try) is one confortable way to do that
-//   if result is some other type and use other aggregate method, like adding numeric value, use only T for that
-//   if result is not care, just use default (unit) for type parameter T and E
-// TODO: use Try when stablized, see branch try-v2 // will I still work on this project at that time?
-pub trait Visitor<T: Default = (), E = ()>: Sized {
-    fn visit_array_expr(&mut self, node: &ArrayExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_array_index_expr(&mut self, node: &ArrayIndexExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_array_type(&mut self, node: &ArrayType) -> Result<T, E> { node.walk(self) }
-    fn visit_assign_expr_stmt(&mut self, node: &AssignExprStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_binary_expr(&mut self, node: &BinaryExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_block(&mut self, node: &Block) -> Result<T, E> { node.walk(self) }
-    fn visit_block_stmt(&mut self, node: &BlockStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_break_stmt(&mut self, node: &BreakStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_call_expr(&mut self, node: &CallExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_cast_segment(&mut self, node: &CastSegment) -> Result<T, E> { node.walk(self) }
-    fn visit_class_def(&mut self, node: &ClassDef) -> Result<T, E> { node.walk(self) }
-    fn visit_continue_stmt(&mut self, node: &ContinueStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_else_clause(&mut self, node: &ElseClause) -> Result<T, E> { node.walk(self) }
-    fn visit_enum_def(&mut self, node: &EnumDef) -> Result<T, E> { node.walk(self) }
-    fn visit_enum_def_variant(&mut self, node: &EnumDefVariant) -> Result<T, E> { node.walk(self) }
-    fn visit_expr(&mut self, node: &Expr) -> Result<T, E> { node.walk(self) }
-    fn visit_field_def(&mut self, node: &FieldDef) -> Result<T, E> { node.walk(self) }
-    fn visit_fn_def(&mut self, node: &FnDef) -> Result<T, E> { node.walk(self) }
-    fn visit_fn_def_parameter(&mut self, node: &FnDefParameter) -> Result<T, E> { node.walk(self) }
-    fn visit_fn_type(&mut self, node: &FnType) -> Result<T, E> { node.walk(self) }
-    fn visit_fn_type_parameter(&mut self, node: &FnTypeParameter) -> Result<T, E> { node.walk(self) }
-    fn visit_for_stmt(&mut self, node: &ForStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_generic_name(&mut self, node: &GenericName) -> Result<T, E> { node.walk(self) }
-    fn visit_generic_parameter(&mut self, node: &GenericParameter) -> Result<T, E> { node.walk(self) }
-    fn visit_generic_segment(&mut self, node: &GenericSegment) -> Result<T, E> { node.walk(self) } 
-    fn visit_if_clause(&mut self, node: &IfClause) -> Result<T, E> { node.walk(self) }
-    fn visit_if_stmt(&mut self, node: &IfStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_impl(&mut self, node: &Implementation) -> Result<T, E> { node.walk(self) }
-    fn visit_item(&mut self, node: &Item) -> Result<T, E> { node.walk(self) }
-    fn visit_lit_expr(&mut self, node: &LitExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_loop_stmt(&mut self, node: &LoopStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_member_expr(&mut self, node: &MemberExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_module(&mut self, node: &Module) -> Result<T, E> { node.walk(self) }
-    fn visit_module_stmt(&mut self, node: &ModuleStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_object_expr(&mut self, node: &ObjectExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_object_expr_field(&mut self, node: &ObjectExprField) -> Result<T, E> { node.walk(self) }
-    fn visit_paren_expr(&mut self, node: &ParenExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_path(&mut self, node: &Path) -> Result<T, E> { node.walk(self) }
-    fn visit_path_segment(&mut self, node: &PathSegment) -> Result<T, E> { node.walk(self) }
-    fn visit_primitive_type(&mut self, node: &PrimitiveType) -> Result<T, E> { node.walk(self) }
-    fn visit_range_both_expr(&mut self, node: &RangeBothExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_range_full_expr(&mut self, node: &RangeFullExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_range_left_expr(&mut self, node: &RangeLeftExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_range_right_expr(&mut self, node: &RangeRightExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_ref_type(&mut self, node: &RefType) -> Result<T, E> { node.walk(self) }
-    fn visit_ret_stmt(&mut self, node: &ReturnStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_simple_expr_stmt(&mut self, node: &SimpleExprStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_simple_segment(&mut self, node: &SimpleSegment) -> Result<T, E> { node.walk(self) }
-    fn visit_stmt(&mut self, node: &Statement) -> Result<T, E> { node.walk(self) }
-    fn visit_struct_def(&mut self, node: &StructDef) -> Result<T, E> { node.walk(self) }
-    fn visit_tuple_expr(&mut self, node: &TupleExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_tuple_index_expr(&mut self, node: &TupleIndexExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_tuple_type(&mut self, node: &TupleType) -> Result<T, E> { node.walk(self) }
-    fn visit_type_def(&mut self, node: &TypeDef) -> Result<T, E> { node.walk(self) }
-    fn visit_type_list(&mut self, node: &TypeList) -> Result<T, E> { node.walk(self) }
-    fn visit_type_ref(&mut self, node: &TypeRef) -> Result<T, E> { node.walk(self) }
-    fn visit_unary_expr(&mut self, node: &UnaryExpr) -> Result<T, E> { node.walk(self) }
-    fn visit_use_stmt(&mut self, node: &UseStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_var_decl(&mut self, node: &VarDeclStatement) -> Result<T, E> { node.walk(self) }
-    fn visit_where_clause(&mut self, node: &WhereClause) -> Result<T, E> { node.walk(self) }
-    fn visit_while_stmt(&mut self, node: &WhileStatement) -> Result<T, E> { node.walk(self) }
+// return type is try for easily break from visiting process if any error happens or result already found, etc.
+
+use std::ops::Try;
+use crate::common::arena::{Arena, Index};
+use super::ast::*;
+
+pub trait Visit: Sized {
+
+    fn walk<V: Visitor>(&self, _arena: &Arena, _visitor: &mut V) -> V::Result {
+        Default::default()
+    }
+
+    fn accept<V: Visitor>(&self, _arena: &Arena, _visitor: &mut V) -> V::Result;
+}
+
+pub trait Visitor: Sized {
+    // it is associate type not generic becuase
+    // 1. it is not very reasonable to implement visitor for one type multiple times
+    // 2. it is very not reasonable to distinguish different implementation by return type
+    type Result: Default + Try;
+
+    // AUTOGEN
+    fn visit_array_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ArrayExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_array_index_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ArrayIndexExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_array_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ArrayType<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_assign_expr_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, AssignExprStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_binary_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, BinaryExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_block<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Block<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_block_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, BlockStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_break_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, BreakStatement>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_call_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, CallExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_cast_segment<'a, 'b: 'a>(&mut self, node: &'b Index<'a, CastSegment<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_class_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ClassDef<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_continue_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ContinueStatement>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_else_clause<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ElseClause<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_enum_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, EnumDef<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_enum_def_variant<'a, 'b: 'a>(&mut self, node: &'b Index<'a, EnumDefVariant<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_expr<'a, 'b: 'a>(&mut self, node: &'b Expr<'a>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_field_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FieldDef<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_fn_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnDef<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_fn_def_parameter<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnDefParameter<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_fn_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnType<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_fn_type_parameter<'a, 'b: 'a>(&mut self, node: &'b Index<'a, FnTypeParameter<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_for_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ForStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_generic_name<'a, 'b: 'a>(&mut self, node: &'b Index<'a, GenericName<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_generic_parameter<'a, 'b: 'a>(&mut self, node: &'b Index<'a, GenericParameter>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_generic_segment<'a, 'b: 'a>(&mut self, node: &'b Index<'a, GenericSegment<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_if_clause<'a, 'b: 'a>(&mut self, node: &'b Index<'a, IfClause<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_if_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, IfStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_impl_block<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Implementation<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_item<'a, 'b: 'a>(&mut self, node: &'b Item<'a>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_lit_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, LitExpr>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_loop_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, LoopStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_member_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, MemberExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_module<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Module<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_module_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ModuleStatement>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_object_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ObjectExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_object_expr_field<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ObjectExprField<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_paren_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ParenExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_path<'a, 'b: 'a>(&mut self, node: &'b Index<'a, Path<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_path_segment<'a, 'b: 'a>(&mut self, node: &'b PathSegment<'a>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_primitive_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, PrimitiveType>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_range_both_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeBothExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_range_full_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeFullExpr>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_range_left_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeLeftExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_range_right_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RangeRightExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_ref_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, RefType<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_ret_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, ReturnStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_simple_expr_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, SimpleExprStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_simple_segment<'a, 'b: 'a>(&mut self, node: &'b Index<'a, SimpleSegment>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_stmt<'a, 'b: 'a>(&mut self, node: &'b Statement<'a>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_struct_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, StructDef<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_tuple_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TupleExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_tuple_index_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TupleIndexExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_tuple_type<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TupleType<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_type_def<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TypeDef<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_type_list<'a, 'b: 'a>(&mut self, node: &'b Index<'a, TypeList<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_type_ref<'a, 'b: 'a>(&mut self, node: &'b TypeRef<'a>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_unary_expr<'a, 'b: 'a>(&mut self, node: &'b Index<'a, UnaryExpr<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_use_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, UseStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_var_decl_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, VarDeclStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_where_clause<'a, 'b: 'a>(&mut self, node: &'b Index<'a, WhereClause<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
+    fn visit_while_stmt<'a, 'b: 'a>(&mut self, node: &'b Index<'a, WhileStatement<'a>>, arena: &'a Arena) -> Self::Result { node.walk(arena, self) }
 }
