+++ tests/ast/hello.stdout
@@ -1,17 +1,13 @@
-total size 900 > 140
-      [item(simple-expr)(1)] 304 (33.78%) > 8 (5.71%)
-   [path-segment(simple)(1)] 224 (24.89%) > 8 (5.71%)
-             [expr(path)(1)] 80 (8.89%) > 8 (5.71%)
-              [expr(lit)(1)] 80 (8.89%) > 8 (5.71%)
-             [expr(call)(1)] 80 (8.89%) > 8 (5.71%)
-              [call-expr(1)] 48 (5.33%) > 28 (20.00%)
-                   [path(1)] 32 (3.56%) > 16 (11.43%)
-               [lit-expr(1)] 32 (3.56%) > 32 (22.86%)
-         [simple-segment(1)] 12 (1.33%) > 12 (8.57%)
-       [simple-expr-stmt(1)] 8 (0.89%) > 12 (8.57%)
+[all(6)] 120
+[lit-expr(1)] 32 (26.67%)
+[call-expr(1)] 32 (26.67%)
+[simple-expr-stmt(1)] 16 (13.33%)
+[path(1)] 16 (13.33%)
+[simple-segment(1)] 12 (10.00%)
+[module(1)] 12 (10.00%)
 module tests/ast/hello.f3
   simple-expr-stmt <4:1-4:23>
     call-expr <4:1-4:22> () <4:8-4:22>
       path <4:1-4:7>
         segment <4:1-4:7> writeln
-      literal str "hello world" <4:9-4:21>
+      lit-expr str "hello world" <4:9-4:21>
