+++ tests/ast/missing.stdout
@@ -1,75 +1,53 @@
-total size 10455 > 1991
-   [path-segment(simple)(7)] 1568 (15.00%) > 56 (2.81%)
-         [item(var-decl)(4)] 1216 (11.63%) > 32 (1.61%)
-             [expr(lit)(10)] 800 (7.65%) > 80 (4.02%)
-    [type-ref(primitive)(6)] 624 (5.97%) > 48 (2.41%)
-              [stmt(ret)(2)] 608 (5.82%) > 16 (0.80%)
-              [lit-expr(10)] 320 (3.06%) > 320 (16.07%)
-           [expr(binary)(4)] 320 (3.06%) > 32 (1.61%)
-           [stmt(struct)(1)] 304 (2.91%) > 8 (0.40%)
-               [stmt(fn)(1)] 304 (2.91%) > 8 (0.40%)
-            [item(while)(1)] 304 (2.91%) > 8 (0.40%)
-              [item(use)(1)] 304 (2.91%) > 8 (0.40%)
-           [item(struct)(1)] 304 (2.91%) > 8 (0.40%)
-             [item(loop)(1)] 304 (2.91%) > 8 (0.40%)
-               [item(if)(1)] 304 (2.91%) > 8 (0.40%)
-               [item(fn)(1)] 304 (2.91%) > 8 (0.40%)
-             [expr(path)(3)] 240 (2.30%) > 24 (1.21%)
-           [generic-name(4)] 208 (1.99%) > 144 (7.23%)
-                   [path(5)] 160 (1.53%) > 80 (4.02%)
-             [expr(call)(2)] 160 (1.53%) > 16 (0.80%)
-                  [block(5)] 160 (1.53%) > 80 (4.02%)
-              [field-def(5)] 140 (1.34%) > 160 (8.04%)
-            [binary-expr(4)] 132 (1.26%) > 100 (5.02%)
-                 [fn-def(2)] 128 (1.22%) > 88 (4.42%)
-          [type-ref(ref)(1)] 104 (0.99%) > 8 (0.40%)
-         [type-ref(path)(1)] 104 (0.99%) > 8 (0.40%)
-              [call-expr(2)] 96 (0.92%) > 56 (2.81%)
-               [var-decl(4)] 84 (0.80%) > 116 (5.83%)
-         [simple-segment(7)] 84 (0.80%) > 84 (4.22%)
-            [expr(unary)(1)] 80 (0.77%) > 8 (0.40%)
-            [expr(tuple)(1)] 80 (0.77%) > 8 (0.40%)
-      [expr(range-right)(1)] 80 (0.77%) > 8 (0.40%)
-       [expr(range-left)(1)] 80 (0.77%) > 8 (0.40%)
-       [expr(range-full)(1)] 80 (0.77%) > 8 (0.40%)
-             [struct-def(2)] 64 (0.61%) > 40 (2.01%)
-         [primitive-type(6)] 54 (0.52%) > 54 (2.71%)
-             [tuple-expr(1)] 32 (0.31%) > 16 (0.80%)
-                [if-stmt(1)] 32 (0.31%) > 24 (1.21%)
-             [unary-expr(1)] 25 (0.24%) > 21 (1.05%)
-             [while-stmt(1)] 20 (0.19%) > 28 (1.41%)
-               [use-stmt(1)] 20 (0.19%) > 24 (1.21%)
-              [loop-stmt(1)] 20 (0.19%) > 24 (1.21%)
-       [fn-def-parameter(1)] 20 (0.19%) > 24 (1.21%)
-               [ret-stmt(2)] 16 (0.15%) > 24 (1.21%)
-               [ref-type(1)] 16 (0.15%) > 12 (0.60%)
-       [range-right-expr(1)] 16 (0.15%) > 12 (0.60%)
-        [range-left-expr(1)] 16 (0.15%) > 12 (0.60%)
-        [range-full-expr(1)] 8 (0.08%) > 8 (0.40%)
-              [if-clause(1)] 8 (0.08%) > 16 (0.80%)
+[all(71)] 1760
+[lit-expr(10)] 320 (18.18%)
+[field-def(5)] 180 (10.23%)
+[var-decl-stmt(4)] 160 (9.09%)
+[generic-name(4)] 144 (8.18%)
+[binary-expr(4)] 144 (8.18%)
+[fn-def(2)] 96 (5.45%)
+[simple-segment(7)] 84 (4.77%)
+[path(5)] 80 (4.55%)
+[block(5)] 80 (4.55%)
+[primitive-type(6)] 72 (4.09%)
+[call-expr(2)] 64 (3.64%)
+[struct-def(2)] 40 (2.27%)
+[while-stmt(1)] 32 (1.82%)
+[ret-stmt(2)] 32 (1.82%)
+[unary-expr(1)] 28 (1.59%)
+[fn-def-parameter(1)] 28 (1.59%)
+[use-stmt(1)] 24 (1.36%)
+[loop-stmt(1)] 24 (1.36%)
+[if-stmt(1)] 24 (1.36%)
+[if-clause(1)] 20 (1.14%)
+[tuple-expr(1)] 16 (0.91%)
+[ref-type(1)] 16 (0.91%)
+[range-right-expr(1)] 16 (0.91%)
+[range-left-expr(1)] 16 (0.91%)
+[module(1)] 12 (0.68%)
+[range-full-expr(1)] 8 (0.45%)
 module tests/ast/missing.f3
-  var-def <3:1-3:16> const full <3:7-3:10>
+  var-decl-stmt <3:1-3:16> const full <3:7-3:10>
     range-full-expr <3:14-3:15>
-  var-def <4:1-4:21> const left <4:7-4:10>
+  var-decl-stmt <4:1-4:21> const left <4:7-4:10>
     range-right-expr <4:14-4:20>
       binary-expr <4:16-4:20> + <4:18-4:18>
-        literal i32 1 <4:16-4:16>
-        literal i32 2 <4:20-4:20>
-  var-def <5:1-5:24> const right <5:7-5:11>
+        lit-expr i32 1 <4:16-4:16>
+        lit-expr i32 2 <4:20-4:20>
+  var-decl-stmt <5:1-5:24> const right <5:7-5:11>
     range-left-expr <5:15-5:22>
       binary-expr <5:15-5:19> * <5:17-5:17>
-        literal i32 3 <5:15-5:15>
-        literal i32 5 <5:19-5:19>
+        lit-expr i32 3 <5:15-5:15>
+        lit-expr i32 5 <5:19-5:19>
   use-stmt <7:1-7:12>
     path <7:5-7:11>
       segment <7:5-7:5> a
       segment <7:8-7:8> b
       segment <7:11-7:11> c
-  var-def <8:1-8:32> mutable some_tuple <8:5-8:14>
+  var-decl-stmt <8:1-8:32> mutable some_tuple <8:5-8:14>
     unary-expr <8:18-8:31> ! <8:18-8:18>
       tuple-expr <8:19-8:31>
-        literal str "hello" <8:20-8:26>
-        literal i32 42 <8:29-8:30>
+        lit-expr str "hello" <8:20-8:26>
+        lit-expr i32 42 <8:29-8:30>
   struct-def <10:1-14:1>
     generic-name <10:8-10:13> string <10:8-10:13>
     field-def <11:5-11:12> len <11:5-11:7> : <11:8-11:8>
@@ -82,13 +60,13 @@ module tests/ast/missing.f3
   if-stmt <16:1-17:1>
     if-clause <16:1-17:1>
       binary-expr <16:4-16:9> != <16:6-16:7>
-        literal i32 1 <16:4-16:4>
-        literal i32 1 <16:9-16:9>
+        lit-expr i32 1 <16:4-16:4>
+        lit-expr i32 1 <16:9-16:9>
       block <16:11-17:1>
   loop-stmt <18:1-19:1>
     block <18:6-19:1>
   while-stmt <20:1-21:1>
-    literal bool true <20:7-20:10>
+    lit-expr bool true <20:7-20:10>
     block <20:12-21:1>
   fn-def <23:1-34:1> () <23:8-23:9>
     generic-name <23:4-23:7> main <23:4-23:7>
@@ -108,7 +86,7 @@ module tests/ast/missing.f3
         block <29:34-31:5>
           ret-stmt <30:9-30:21>
             binary-expr <30:16-30:20> * <30:18-30:18>
-              literal i32 2 <30:16-30:16>
+              lit-expr i32 2 <30:16-30:16>
               path <30:20-30:20>
                 segment <30:20-30:20> a
       ret-stmt <33:5-33:34>
